ICD class refactoring
