--- src/main/java/peanut/medicine/ICDReader/ICDinputfilter.java
@@ -1,28 +1,0 @@
-package peanut.medicine.ICDReader;
-
-/**
- * Created by Matheo on 16.03.2017.
- */
-public class ICDinputfilter {
-
-    private String inputString;
-
-    public ICDinputfilter(String inputString) {
-        this.inputString = inputString;
-    }
-
-    public String getInputString() {
-        return inputString;
-    }
-
-    public void setInputString(String inputString) {
-        this.inputString = inputString;
-    }
-
-    @Override
-    public String toString() {
-        return super.toString();
-    }
-
-
-}
